# public/swagger.json
{
	"components": {
...
		"schemas": {
			"PlayerLevel": {
				"enum": [
					"NEWBIE",
					"BEGINNER",
					"INTERMEDIATE",
					"ADVANCED",
					"PRO"
				],
				"type": "string"
			},
...
			"PlayersResponse": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"nickname": {
						"type": "string"
					},
					"level": {
						"$ref": "#/components/schemas/PlayerLevel"
					},
					"address": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"country": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"nickname",
					"level",
					"address",
					"email",
					"city",
					"country"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UpdatePlayerRequest": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"level": {
						"$ref": "#/components/schemas/PlayerLevel"
					},
					"address": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"country": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"ClubCreateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"country": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"address",
					"description",
					"city",
					"country"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CourtSurface": {
				"enum": [
					"GRASS",
					"CLAY",
					"HARD"
				],
				"type": "string"
			},
			"ClubResponse": {
				"properties": {
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"courts": {
						"items": {
							"$ref": "#/components/schemas/CourtResponse"
						},
						"type": "array"
					},
					"courtsNumber": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"address",
					"description",
					"city",
					"country",
					"courtsNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CourtResponse": {
				"properties": {
					"name": {
						"type": "string"
					},
					"surface": {
						"$ref": "#/components/schemas/CourtSurface"
					},
					"stands": {
						"type": "boolean"
					},
					"roof": {
						"type": "boolean"
					},
					"double": {
						"type": "boolean"
					},
					"clubId": {
						"type": "string"
					},
					"club": {
						"$ref": "#/components/schemas/ClubResponse"
					},
					"pricePerHour": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"surface",
					"stands",
					"roof",
					"double",
					"clubId",
					"pricePerHour"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateClubRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"courts": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"address",
					"description",
					"city",
					"country",
					"courts"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"BookingType": {
				"enum": [
					"ONE_TIME",
					"MONTH",
					"SEASON"
				],
				"type": "string"
			},
			"BookingCreateRequest": {
				"properties": {
					"court": {
						"type": "string"
					},
					"from": {
						"type": "number",
						"format": "double"
					},
					"to": {
						"type": "number",
						"format": "double"
					},
					"totalPrice": {
						"type": "number",
						"format": "double"
					},
					"player": {
						"type": "string"
					},
					"bookingType": {
						"$ref": "#/components/schemas/BookingType"
					},
					"date": {
						"type": "string"
					}
				},
				"required": [
					"court",
					"from",
					"to",
					"player",
					"bookingType",
					"date"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BookingResponse": {
				"properties": {
					"entityId": {
						"type": "string"
					},
					"clubName": {
						"type": "string"
					},
					"clubAddress": {
						"type": "string"
					},
					"clubCity": {
						"type": "string"
					},
					"courtName": {
						"type": "string"
					},
					"courtSurface": {
						"type": "string"
					},
					"from": {
						"type": "number",
						"format": "double"
					},
					"to": {
						"type": "number",
						"format": "double"
					},
					"totalPrice": {
						"type": "number",
						"format": "double"
					},
					"playerFirstName": {
						"type": "string"
					},
					"playerLastName": {
						"type": "string"
					},
					"bookingType": {
						"type": "string"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"entityId",
					"clubName",
					"clubAddress",
					"clubCity",
					"courtName",
					"courtSurface",
					"from",
					"to",
					"totalPrice",
					"playerFirstName",
					"playerLastName",
					"bookingType",
					"date"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateBookingRequest": {
				"properties": {
					"court": {
						"type": "string"
					},
					"from": {
						"type": "number",
						"format": "double"
					},
					"to": {
						"type": "number",
						"format": "double"
					},
					"totalPrice": {
						"type": "number",
						"format": "double"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"court",
					"from",
					"to",
					"totalPrice",
					"date"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "tennis",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/players/{entityId}": {
			"get": {
				"operationId": "GetByEntityId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlayersResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeletePlayer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdatePlayer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePlayerRequest"
							}
						}
					}
				}
			}
		},
...
		"/clubs": {
			"post": {
				"operationId": "CreateClub",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Clubs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ClubCreateRequest"
							}
						}
					}
				}
			}
		},
		"/clubs/all": {
			"get": {
				"operationId": "GetAllClubs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ClubResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Clubs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/clubs/city/{city}": {
			"get": {
				"operationId": "GetClubsByCity",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ClubResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Clubs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "city",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/clubs/{entityId}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ClubResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Clubs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteClub",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Clubs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateClub",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Clubs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateClubRequest"
							}
						}
					}
				}
			}
		},
...
		"/bookings": {
			"post": {
				"operationId": "CreateBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BookingCreateRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FilterBookings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/BookingResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "court",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "from",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "player",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "date",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/bookings/{entityId}": {
			"delete": {
				"operationId": "DeleteBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBookingRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetByEntityId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookingResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}